add snippet to modal
